requirements.txt
pydantic-ai[mcp,web,openai]
sap-ai-sdk-gen[all]
mcp
httpx
uvicorn
python-dotenv
fastapi
jinja2
python-multipart
sqlalchemy[asyncio]>=2.0
asyncpg
pyjwt[crypto]
